adjust text-layer for new events
